requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
google-generativeai==0.8.3
openai>=1.57.0
aiohttp==3.10.11
aiosqlite==0.19.0
httpx>=0.27.0
